+++ lib/Braintree/CreditCardVerification.php
@@ -18,6 +18,7 @@ namespace Braintree;
  * @property-read \DateTime $createdAt
  * @property-read \Braintree\CreditCard $creditCard
  * @property-read string|null $gatewayRejectionReason
+ * @property-read string $graphQLId
  * @property-read string $id
  * @property-read string $merchantAccountId
  * @property-read string $processorResponseCode