+++ lib/Braintree/Result/CreditCardVerification.php
@@ -62,6 +62,12 @@ class CreditCardVerification
             $attributes['riskData'] = RiskData::factory($attributes['riskData']);
         }
 
+        // TODO investigate why CC Verifications don't inherit from the Base result class.
+        if(isset($attributes['globalId']))
+        {
+            $attributes['graphQLId'] = $attributes['globalId'];
+        }
+
         if(isset($attributes['threeDSecureInfo'])) {
             $attributes['threeDSecureInfo'] = ThreeDSecureInfo::factory($attributes['threeDSecureInfo']);
         }