+++ lib/Braintree/Transaction/AndroidPayCardDetails.php
@@ -28,6 +28,7 @@ use Braintree\Instance;
  * @property-read string $googleTransactionId
  * @property-read string $healthcare
  * @property-read string $imageUrl
+ * @property-read boolean $isNetworkTokenized
  * @property-read string $issuingBank
  * @property-read string $payroll
  * @property-read string $prepaid