+++ tests/integration/TransactionTest.php
@@ -1752,6 +1752,7 @@ class TransactionTest extends Setup
         $this->assertNotNull($androidPayCardDetails->payroll);
         $this->assertNotNull($androidPayCardDetails->prepaid);
         $this->assertNotNull($androidPayCardDetails->productId);
+        $this->assertFalse($androidPayCardDetails->isNetworkTokenized);
     }
 
   public function testCreateTransactionUsingFakeAndroidPayNetworkTokenNonce()
@@ -1784,6 +1785,7 @@ class TransactionTest extends Setup
         $this->assertNotNull($androidPayCardDetails->payroll);
         $this->assertNotNull($androidPayCardDetails->prepaid);
         $this->assertNotNull($androidPayCardDetails->productId);
+        $this->assertTrue($androidPayCardDetails->isNetworkTokenized);
     }
 
     public function testCreateTransactionUsingFakeAmexExpressCheckoutNonce()
@@ -3247,6 +3249,7 @@ class TransactionTest extends Setup
         $this->assertEquals('100.00', $transaction->amount);
         $this->assertEquals('510510', $transaction->creditCardDetails->bin);
         $this->assertEquals('5100', $transaction->creditCardDetails->last4);
+        $this->assertNotNull($transaction->graphQLId);
     }
 
   public function testFindExposesDisbursementDetails()